requirements.txt
av==13.1.0
bitsandbytes==0.42.0
datasets==2.17.0
decord==0.6.0
deepspeed==0.14.4
flash_attn==2.6.2
fsspec==2023.10.0
fvcore==0.1.5.post20221221
huggingface_hub==0.24.6
matplotlib==3.7.1
natsort==8.4.0
numpy==1.24.3
openai==1.55.0
packaging==24.2
pandas==2.0.0
peft==0.11.1
Pillow==11.0.0
pycocoevalcap==1.2
pycocotools==2.0.6
python_magic==0.4.27
PyYAML==6.0.2
requests==2.32.3
shortuuid==1.0.13
thop==0.1.1.post2209072238
timm==0.6.13
tokenizers==0.20.3
torch==2.5.0
torchvision==0.16.2
tqdm==4.65.0
transformers==4.45.2
